Revert "Database-setup"
This reverts commit a62fa9c0da36ab19ef57c505716845eee103de79.

+++ backend/src/main/java/com/eku/eku_ocr_test/form/SignUpForm.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class SignUpForm {
-    private Long studNo;
+    private int studNo;
     private String name;
     private String department;
     private String email;
